+++ src/collection.spec.ts
@@ -58,6 +58,12 @@ describe("Collection", () => {
                 nameColumn: 'Name',
                 countColumn: 'Count',
                 syncIntervalMs: 10
+            },
+            decklist: {
+                preferredCurrency: 'usd',
+                showCardNamesAsHyperlinks: true,
+                showCardPreviews: true,
+                showBuylist: true
             }
         }
         test('handles multiple CSVs', async () => {