+++ src/scryfall.spec.ts
@@ -1,0 +1,40 @@
+import {describe, expect, test, jest } from '@jest/globals';
+import { RequestOptions } from 'https';
+import { getCardData, getMultipleCardData } from './scryfall';
+import { EXAMPLE_MULTI_CARD_RESPONSE, EXAMPLE_SCRYFALL_RESPONSE_1 } from '../jest/fixtures/scryfall-data';
+
+describe("Scryfall", () => {
+    describe("#getCardData()", () => {
+        test("for a single card", async () => {
+            function httpReq<ScryfallResponse>(options: RequestOptions): Promise<ScryfallResponse> {
+                return new Promise((res) => {
+                    res(EXAMPLE_SCRYFALL_RESPONSE_1 as unknown as ScryfallResponse);
+                })
+            };
+            const data = await getCardData(
+                "Delver of Secrets", 
+                httpReq
+            );
+            expect(data).toEqual(EXAMPLE_SCRYFALL_RESPONSE_1);
+        });
+    });
+
+    describe("#getMultipleCardData()", () => {
+        test("for a single card", async () => {
+            function httpReq<ScryfallResponse>(options: RequestOptions, postBody: string): Promise<ScryfallResponse> {
+                return new Promise((res) => {
+                    res(EXAMPLE_MULTI_CARD_RESPONSE as unknown as ScryfallResponse);
+                })
+            };
+            const data = await getMultipleCardData(
+                [
+                    "Delver of Secrets",
+                    "Ledger Shredder",
+                    "Dark Ritual"
+                ], 
+                httpReq
+            );
+            expect(data).toEqual(EXAMPLE_MULTI_CARD_RESPONSE);
+        });
+    });
+});