+++ src/settings.ts
@@ -8,5 +8,15 @@ export interface ObsidianPluginMtgSettings {
 		countColumn: string;
         // The number of milliseconds between collection count syncs
         syncIntervalMs: number;
+	};
+	decklist: {
+		// Card Price Currency:
+		preferredCurrency: 'usd' | 'eur' | 'tix';
+		// Show hyperlinks
+		showCardNamesAsHyperlinks: boolean;
+		// Show card previews
+		showCardPreviews: boolean;
+		// Show buylist
+		showBuylist: boolean;
 	}
 }